Updates
implemented HTML and CSS code for graph, made the countdown part work for arduino 172.20.10.10 but countodwn doesnt work on index.html because of fetch() not working between the server and index.html.

as for now the bar graph is static as i am still not able to make font and logo appear on server site. also visuals.html and index.html links dont work on server site yet, have not checked this out.

will take a break from this project for now.